Accept filehandles in trace_calls()
This changes adds an alternative to logging
to STDERR or a file name "$file.$PID".
Either globs, globrefs or filehandles
can be given as valid targets for logging.

Some remarks to be noticed:
(1) all handles (including STDERR) are
    duped with Perl "open $fh, '>>&'

(2) for "$file.$PID", now a 3-args open
    is used

(3) An additional dependency to Scalar::Util
    is introduced